Separate package for tests added
