+++ Checkers-master/nbproject/private/private.properties
@@ -1,2 +1,2 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\Hashini\\AppData\\Roaming\\NetBeans\\8.0\\build.properties
+user.properties.file=C:\\Users\\Dinu\\AppData\\Roaming\\NetBeans\\7.3\\build.properties
