+++ Checkers-master/nbproject/private/private.xml
@@ -1,11 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/1">
-        <file>file:/C:/Users/user/Documents/GitHub/Checkers/src/Logic/GameEngine.java</file>
-        <file>file:/C:/Users/user/Documents/GitHub/Checkers/src/Logic/HeuristicUtil.java</file>
-        <file>file:/C:/Users/user/Documents/GitHub/Checkers/src/Logic/CheckerBoard.java</file>
-    </open-files>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
             <file>file:/F:/sem5/CS3612%20Intelligent%20Systems/project/Latest/Checkers/src/Logic/GameEngine.java</file>
