--- Checkers-master/src/Logic/CheckerBoardTest.java
@@ -1,184 +1,0 @@
-package Logic;
-
-import static org.junit.Assert.fail;
-
-import org.junit.Test;
-
-public class CheckerBoardTest {
-
-  @Test
-  public void testCheckerBoard() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testPrintBoard() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testMovePiece() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsMoveable() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testCutPiece() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testPieceCount() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsUsed() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsUsedByNormalPiece() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsUsedByQueen() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testGetSize() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testGetTypeRList() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testGetTypeBList() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testCalcHeuristicValue() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testCalcHeuristicValueEnding() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testGetTypeR() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testGetTypeB() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsMoveableByType() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testMovePieceByType() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testHasMovesIntInt() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testHasMovesChar() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testMove() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testHasCutsIntInt() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testHasCutsCharArrayArrayIntInt() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testHasMoreCuts() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testCutPieceByType() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testCanMove() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testGetAllCaptures() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testGetAllNonCaptures() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testCloneArray() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testGetMoveFromMinMax() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testInvertColour() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testFindValue() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testCalcHeuristic() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testEval() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testReset() {
-    fail("Not yet implemented");
-  }
-
-}
