--- Checkers-master/src/Logic/HeuristicUtilTest.java
@@ -1,44 +1,0 @@
-package Logic;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-public class HeuristicUtilTest {
-
-  @Test
-  public void testCalcHeuristicValue() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testCalcHeuristicValueEnding() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsUsedByNormalPiece() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsUsedByQueen() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsUsed() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testPieceCount() {
-    fail("Not yet implemented");
-  }
-
-  @Test
-  public void testIsOnlyKings() {
-    fail("Not yet implemented");
-  }
-
-}
